backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
requests
python-dateutil
